symptom/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
google-generativeai==0.3.2
googletrans==4.0.0-rc1
python-dotenv==1.0.0
gTTS==2.3.2
